+++ src/main/java/pkg/reply/ReplyDAOImpl.java
@@ -1,0 +1,38 @@
+package pkg.reply;
+
+import org.mybatis.spring.SqlSessionTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+@Repository
+public class ReplyDAOImpl implements ReplyDAO{
+	
+	SqlSessionTemplate sqlSession;
+	
+	@Autowired
+	public ReplyDAOImpl(SqlSessionTemplate sqlSession){
+		this.sqlSession = sqlSession;
+	}
+	
+	public void addReply(){
+	
+	}
+	
+	public List<ReplyVO> selectReplyList(String replyArticleId){
+		Map<String, Object> map = new HashMap<>();
+		map.put("replyArticleId",replyArticleId);
+		
+		sqlSession.selectList("selectReplyList",map);
+		
+		List<ReplyVO> resultList = (List<ReplyVO>) map.get("result");
+		
+		return resultList;
+	}
+	
+	
+}
+
