+++ webapp/WEB-INF/dispatcher-servlet.xml
@@ -35,21 +35,21 @@
    </bean>
    -->
     <!-- 유승화 db연결 -->
-    <!--
+
     <bean id="dataSource" class="org.springframework.jdbc.datasource.DriverManagerDataSource">
       <property name="driverClassName" value="oracle.jdbc.driver.OracleDriver"/>
       <property name="url" value="jdbc:oracle:thin:@DB20220621092517_high?TNS_ADMIN=/Users/sniperditto/Documents/Project/Wallet_DB20220621092517"/>
       <property name="username" value="FINALPROJECT"/>
       <property name="password" value="Project123123"/>
    </bean>
-    -->
+    <!--
     <bean id="dataSource" class="org.springframework.jdbc.datasource.DriverManagerDataSource">
       <property name="driverClassName" value="oracle.jdbc.driver.OracleDriver"/>
       <property name="url" value="jdbc:oracle:thin:@DB20220621092517_high?TNS_ADMIN=\\or\\Wallet_DB20220621092517"/>
       <property name="username" value="FINALPROJECT"/>
       <property name="password" value="Project123123"/>
      </bean>
-     
+     -->
     
    <bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
       <property name="dataSource" ref="dataSource" />
